% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_or.R
\name{stratified_or}
\alias{stratified_or}
\title{Crude and Stratified Odds Ratio (OR) Estimation with 95% Confidence Interval}
\usage{
stratified_or(exposure, outcome, confounder)
}
\arguments{
\item{exposure}{A numeric or factor vector. The exposure may contain multiple levels. The lowest level is treated as the reference group.}

\item{outcome}{A numeric or factor vector. The outcome can only contain two levels. The lowest level is treated as the reference group.}

\item{confounder}{A vector representing a potential confounder used for stratification. Each unique non-missing value defines a stratum.}
}
\value{
A list containing two components:
 \describe{
 \item{Crude}{A list containing the crude contingency table and the crude odds ratio with its 95\% confidence interval.}
 \item{Stratified}{A named list of stratum-specific results. For each stratum, the output includes a contingency table and the corresponding odds ratio with its 95\% confidence interval.}
 }
}
\description{
Computes the crude odds ratio (OR) and stratum-specific odds ratios by a specified confounder using raw exposure and outcome vectors. This function automates the construction of two-way contingency tables with totals and provides odds ratio estimates with 95% confidence intervals.
}
\details{
Crude and stratum-specific two-way frequency tables with totals are presented.
Odds ratios are estimated using the standard maximum likelihood estimator based on 2*2 contingency tables.
Confidence intervals are computed using the mid-p exact method as the default computation method implemented by \code{\link[epitools:oddsratio]{oddsratio()}}
Missing values from exposure, outcome and confounder variables are excluded from the analysis. Counts of missing values are returned.
If any cell in a contingency table contains less than 5 counts, a warning message is returned as implemented by \code{\link[epitools:oddsratio]{oddsratio()}}
}
\examples{
## Generate simulation data
set.seed(123)
exposure <- rbinom(100, 1, 0.4)
outcome <- rbinom(100, 1, 0.3)
confounder <- sample(1:3, 100, replace = TRUE)
## Run stratified odds ratio analysis
stratified_or(exposure, outcome, confounder)
}
